# expense-management/openapi.yaml
openapi: 3.1.0
info:
  title: 経費精算システム API
  description: |
    中小〜中規模企業向け経費精算システムのREST API仕様。
    マルチテナント対応、RBAC認可、多段階承認フローを実装。
  version: 1.0.0
  contact:
    name: API Support
    email: support@expense.example.com
servers:
  - url: https://api.expense.example.com/api/v1
    description: 本番環境
  - url: https://staging-api.expense.example.com/api/v1
    description: ステージング環境
  - url: http://localhost:8000/api/v1
    description: 開発環境
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Sanctum
...
  schemas:
...
    ExpenseStatus:
      type: string
      enum:
        - draft
        - pending
        - partially_approved
        - approved
        - rejected
        - cancelled
        - paid
      description: |
        - draft: 下書き
        - pending: 申請中
        - partially_approved: 一部承認
        - approved: 承認済み
        - rejected: 却下
        - cancelled: 取り消し
        - paid: 支払済み
    CreateExpenseRequest:
      type: object
      required: [title, items]
      properties:
        title:
          type: string
          maxLength: 200
          example: "大阪出張交通費"
        description:
          type: string
          maxLength: 2000
          nullable: true
          example: "2024年1月15日 東京-大阪往復"
        currency:
          type: string
          enum: [JPY, USD, EUR]
          default: JPY
        items:
          type: array
          minItems: 1
          maxItems: 50
          items:
            $ref: '#/components/schemas/ExpenseItemInput'
    ExpenseItemInput:
      type: object
      required: [category_id, description, amount, expense_date]
      properties:
        category_id:
          type: string
          format: uuid
        description:
          type: string
          maxLength: 500
          example: "新幹線 東京→大阪"
        amount:
          type: integer
          minimum: 1
          maximum: 99999999
          example: 14060
        quantity:
          type: integer
          minimum: 1
          default: 1
        expense_date:
          type: string
          format: date
          example: "2024-01-15"
        vendor:
          type: string
          maxLength: 200
          nullable: true
          example: "JR東海"
        purpose:
          type: string
          maxLength: 500
          nullable: true
          example: "顧客訪問"
    Expense:
      type: object
      properties:
        id:
          type: string
          format: uuid
        expense_number:
          type: string
          example: "EXP-2401-000001"
        title:
          type: string
        description:
          type: string
          nullable: true
        total_amount:
          type: integer
          example: 28120
        total_amount_formatted:
          type: string
          example: "¥28,120"
        currency:
          type: string
          example: "JPY"
        status:
          $ref: '#/components/schemas/ExpenseStatus'
        status_label:
          type: string
          example: "申請中"
        current_step:
          type: integer
          example: 1
        applied_at:
          type: string
          format: date-time
          nullable: true
        approved_at:
          type: string
          format: date-time
          nullable: true
        paid_at:
          type: string
          format: date-time
          nullable: true
        applicant:
          $ref: '#/components/schemas/UserSummary'
        items:
          type: array
          items:
            $ref: '#/components/schemas/ExpenseItem'
        can_edit:
          type: boolean
        can_submit:
          type: boolean
        can_cancel:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ExpenseItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        category_id:
          type: string
          format: uuid
        category:
          $ref: '#/components/schemas/CategorySummary'
        description:
          type: string
        amount:
          type: integer
        quantity:
          type: integer
        line_total:
          type: integer
        line_total_formatted:
          type: string
        expense_date:
          type: string
          format: date
        vendor:
          type: string
          nullable: true
        purpose:
          type: string
          nullable: true
        receipts:
          type: array
          items:
            $ref: '#/components/schemas/Receipt'
    Receipt:
      type: object
      properties:
        id:
          type: string
          format: uuid
        original_filename:
          type: string
        mime_type:
          type: string
        file_size:
          type: integer
        file_size_formatted:
          type: string
          example: "2.4 MB"
        url:
          type: string
          description: "S3署名付きURL（60分有効）"
        ocr:
          type: object
          nullable: true
          properties:
            text:
              type: string
            amount:
              type: integer
            date:
              type: string
              format: date
            vendor:
              type: string
        created_at:
          type: string
          format: date-time
...
    UserSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
        department:
          type: string
          nullable: true
    CategorySummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        code:
          type: string
          nullable: true
...
    PaginatedMeta:
      type: object
      properties:
        total:
          type: integer
        per_page:
          type: integer
        current_page:
          type: integer
        last_page:
          type: integer
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
security:
  - bearerAuth: []
paths:
...
  /expenses:
    get:
      tags: [Expenses]
      summary: 経費申請一覧取得
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/ExpenseStatus'
        - name: applicant_id
          in: query
          schema:
            type: string
            format: uuid
        - name: date_from
          in: query
          schema:
            type: string
            format: date
        - name: date_to
          in: query
          schema:
            type: string
            format: date
        - name: amount_min
          in: query
          schema:
            type: integer
        - name: amount_max
          in: query
          schema:
            type: integer
        - name: keyword
          in: query
          schema:
            type: string
          description: タイトル・説明の全文検索
        - name: per_page
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: sort_by
          in: query
          schema:
            type: string
            enum: [created_at, applied_at, total_amount, expense_number]
            default: created_at
        - name: sort_dir
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: 一覧取得成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Expense'
                  meta:
                    $ref: '#/components/schemas/PaginatedMeta'

    post:
      tags: [Expenses]
      summary: 経費申請作成（下書き）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateExpenseRequest'
            example:
              title: "大阪出張交通費"
              description: "2024年1月15日 東京-大阪往復"
              currency: "JPY"
              items:
                - category_id: "uuid-here"
                  description: "新幹線 東京→大阪"
                  amount: 14060
                  expense_date: "2024-01-15"
                  vendor: "JR東海"
                  purpose: "顧客訪問"
      responses:
        '201':
          description: 作成成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Expense'
        '422':
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...